bench: organize benchmarks by sketch workload
