+++ Cargo.toml
@@ -16,7 +16,7 @@
 # under the License.
 
 [workspace]
-members = ["datasketches", "tests-integration", "xtask"]
+members = ["benchmarks", "datasketches", "tests-integration", "xtask"]
 resolver = "3"
 
 [workspace.package]
